Gentle control of dimmers
Added a bunch of code to support "gently" setting dimmers to new dim levels.  

Many physical dimmers can't dim to anything other than their last dim value.  And setLevel() causes the dimmers to turn of, if they are off.  But changing dim levels in an occupied room, or illuminating a currently-dark room is not always desirable.

A "gentle" dimmer setting therefore is one where:
 - Nothing happens if the light is on at the time of the mode change
 - If/when the dimmer is turned off (whether at the mode change or at a subsequent "off" event),
   the new dim level is asserted using setLevel followed by an "immediate" return to off.  This causes a small
  visible flicker, but it's the best known approach to solving that problem.